requirements.txt
kafka-python==2.0.5
requests==2.32.3
pyspark==3.5.5
numpy==2.0.2